+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 For the source code of each version, see the [tags page](https://github.com/njshockey/keysmith-rs/tags).
 
+# v0.3.0
+***CAUTION***: v0.3.0 introduces breaking changes. Mainly with importing modules. Be careful when upgrading from an older version.
+
+## Features
+- Can now get a timestamp for the current time.
+  - Use get_timestamp() or get_timestamp_i64() for this.
+  - Based on number of seconds since the first second of 2022.
+- Can now generate keys with only numbers and letters.
+  - Use gen_nums_and_letters() for this.
+  - gen_nums_and_letters_lower() for just lowercase letters.
+  - gen_nums_and_letters_upper() for just uppercase letters.
+- Generate a single character with the gen_char module.
+  - gen_char() to generate a character found in gen_key().
+  - gen_uuid_char() to generate a character found in gen_uuid().
+
+## API
+- Functions have been separated into different modules. Import relevant features with keysmith::{key, timestamp, uuid, gen_char} where necessary. See [docs](https://docs.rs/keysmith/latest/keysmith/).
+
+## Docs
+- Rust generated auto docs. Link to newest version can be found on [docs.rs](https://docs.rs/keysmith/latest/keysmith/).
+
+## Testing
+- Can now toggle individual modules and functions.
+- Cleaned up ks-tests crate.
+
+## Development
+- gen_char() and gen_uuid_char() are now public and can be used with the keysmith crate.
+- Cleaned up a lot of code under the hood.
+
 # v0.2.1
 You can now generate keys with only specific kinds of characters. For more details, see [usage](README.md#usage).
 
@@ -18,6 +47,7 @@ You can now generate keys with only specific kinds of characters. For more detai
 
 # v0.1.1
 - gen_key() can now generate uppercase letters.
+- Fixed a build error on Linux.
 
 # v0.1.0
 - First release.